// site/src/generated/BenchmarkPoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BenchmarkPoint = { 
/**
 * Input size (x axis, plotted on a log scale).
 */
input_size: number, 
/**
 * Median wall time of the kura Rust implementation, in milliseconds.
 */
rust_ms: number, 
/**
 * Median wall time of the reference implementation, in milliseconds.
 */
reference_ms: number, };